hello-libs: Fix dependencies issue with new Gradle plugin
Before Gradle plugin 3.0.0 (Android Studio before 3.0), building app
module was starting building of externalNativeBuild of gen-libs module,
as dependency. Indeed, we have this dependency declared at
app/build.gradle file:

    dependencies {
        compile project(':gen-libs')
    }

But after updating Android Studio to v3.0.1 (and Gradle plugin to v3.0.1)
build error started to appear. That build error is connected with unmet
dependencies (attempt to build libhello-libs.so before gen-libs module
libraries). So I presume it's some bug in new Gradle or new Gradle
plugin. If so, this patch is just a workaround. Still we need it until
the issue is fixed in Gradle/plugin. This issue was first reported as
 #450 (see [1]), and then reported to Google ([2]).

With this patch, next sequence works fine:

  1. Remove pre-built libraries (distribution/ directory)
  2. Enable gen-libs module building (comment out corresponding lines
     in app/build.gradle and settings.gradle files)
  3. Start project building

[1] https://github.com/googlesamples/android-ndk/issues/450
[2] https://issuetracker.google.com/issues/69616088

Signed-off-by: Sam Protsenko <joe.skb7@gmail.com>