Drop bug in workaround for r21+.
This workaround for a bug in abspath used to workaround path length
issues in Windows is not needed in r21+ because the abspath bug was
fixed. Unfortunately the bug being fixed means that the workaround now
introduces a bug when forming the path. Only use the workaround for
pre-r21 NDKs.

Bug: http://b/142666190