Hello-Vulkan 1.0 (#869)
We want to advocate for Vulkan as a preferable alternative to OpenGL in
the Android ecosystem. An initial step is to create a similar
application to the already existing Hello-GL. This is to allow
interested developers to explore Vulkan on Android.

This PR intends to add such sample.

Main features of the sample:

- draws a triangle on a SurfaceView via GameActivity (thanks @ggfan)
using VK_PRESENT_MODE_FIFO_KHR (vulkan's hard V-SYNC option)
- clears the screen colour and slowly flashes from white to black
- enables rotation using Vulkan pre-rotation
- implements Vulkan validation layers (with steps provided in the README
regarding ways to enable them).