Sanitizer: Set useLegacyPackaging for asan
[Why]
Property useLegacyPackaging determin whether to use the legacy convention of compressing all .so files in the APK.
If null, .so files will be uncompressed and page-aligned when minSdk >= 23.
Otherwise, it failed with "INSTALL_FAILED_INVALID_APK: Failed to extract native libraries, res=-2" when minSdk >= 23.
And wrap.sh is only available for API level 27 and above.

[How]
Set useLegacyPackaging to true.
Set compileSdk to 27.
Reference: https://developer.android.com/ndk/guides/asan