Use setup-android to make sdkmanager available.
From the action's docs I don't think sdkmanager is actually unavailable,
just obnoxious to find (seems the SDK moved that component to
`cmdline-tools/$VERSION/bin`, and I've got no idea how to guess the
version).

We could make broader use of this by moving the guts of `setup_env.sh`
into build.yml, but to do that we'd have to remove the auto-detection of
which NDK versions we need (though that's a bug anyway, AGP's supposed
to be installing those for us). For now, just unbreak CI and we can deal
with that later if it's ever worth doing.

Using setup-android also requires that we update to a newer version of
Java in setup-java.