Hello HLSL
Create a Hello HLSL sample app based on the Hello Vulkan project.

HLSL is now supported as a first class Vulkan shading language, and is
well-supported in Android for shaders compiled with DXC. This sample
application is intended to serve as an example, particularly for the
gradle build configuration, necessary to use HLSL shaders in Android
Studio development with the Android NDK and LunarG Vulkan SDK.

To ease code review, this PR is split into 2 commits:
- one that copies hello-vulkan to hello-hlsl
- another that makes the necessary modifications to replace GLSL shaders
  with HLSL
These will be squashed before merging.