Prepare Hello GL for multisampling demo
Adds a few images and an expensive render pass to draw into them.  A
part of the image is blit into the swapchain for visual verification.
This preparation step does the above in single-sampled rendering, with
follow up changes enabling multisampling.

The goal is to demonstrate how multisampling can be done efficiently
(by comparing with single-sampled rendering in this commit) on tiling
GPUs.