Add a global .clang-format.
This can't be enforced automatically because there's no good way to
ensure that every developer is using the same clang-format, and there
are style differences across versions (that's what bug fixes and
improvements do in an auto-formatter). Still it's useful to provide one
of these that can at least be used for reformatting modified sections,
even if it'd cause a lot of churn to apply globally.