Replace `ALooper_pollAll` with `ALooper_pollOnce`.
The `pollAll` API was removed because it can't be used safely. The
recommendation is to replace existing calls to `pollAll` with `pollOnce`
with an explicit loop.

There are quite a few other samples that are doing this. Fixing one to
start so I can check with someone that understands the `pollAll` bug so
I don't accidentally copy the same bug, then I'll fix the others.

https://github.com/android/ndk-samples/issues/999