Move sensor handling to a callback.
The apparent change in `minSdkVersion` here isn't actually changing
much. The NDK hasn't supported ICS in a long time, and 21 is the minimum
for modern NDKs (though I think this sample is still using an NDK that
supports 19, it won't soon). The raise is needed for
android_set_abort_message.