Migrate most samples off ALooper_pollAll.
The only one left is the vulkan sample. That one crashed when I did the
trivial thing, so I'm leaving it for last.

There's a fair amount of absurd stuff left around the code I touched.
Some samples have a lot of code dedicated to managing sensors that then
do nothing with the sensors, some display a static image but have a
render "loop", all of these have a nearly identical main loop but don't
share any code, etc. I'll be back.