Add top level Gradle project.
This is step on of https://github.com/android/ndk-samples/issues/1010.
It's not really finished until I've removed all the stuff left behind
(which I'm leaving for another PR because it's just a big mess of
deletes that'd make this PR harder to read).

It's may be good to also remove the mostly unnecessary middle
directories from most of these. Doing that will break a bunch of
existing URLs, but if the goal is to be a project that's actually
sensible to copy, we need to rearrange to do that. Maybe that's better
left for a broader rearranging that's probably overdue anyway, or maybe
we should just leave it as-is to keep the URLs alive.