Remove cruft from samples build.gradle files.
This has been cargo culted forward in each sample for who knows how
long. Most of these don't use kotlin or the library plugin.

The clean task bit I think was a workaround for a very old AGP bug that
was in the new app templates, but if I create a new project with AS now,
it's no longer included.