Downgrade Kotlin in hello-vulkan to match others.
The Kotlin use here is trivial (it might actually be better to make all
the NativeActivity-esque samples Java, since it's fewer dependencies and
less stuff in the APK, which I expect games typically want). I'll
upgrade the whole repo to something actually modern once I can do it as
a batch as opposed to repo by repo. I need all the kotlin versions
matching before I can do that though, and everything else is on 1.7.21.