audio-echo: remove proguard stuff.
minifyEnabled already defaults to false. aiui that's in all the
templates (and it still is in the template for a newly created project)
as a hint for something developers may want to turn on later. It doesn't
add anything to the sample if we're not actually going to enable it, and
I'm not going to go figure out what proguard changes need to happen to
enable it right now. Maybe later.

Since minification isn't used, the proguard config does nothing.