Move Java/Kotlin compat to the convention plugin.
This was apparently only ever in the gradle scripts for projects that
used Kotlin. I don't know if it's necessary for Java projects, but it's
likely not harmful? For Kotlin projects these versions all need to
match.

It's actually possible for us to bump this to Java 11, by the look of
things (see the Now In Android link below). For now I'm keeping things
the way they are.

https://github.com/android/nowinandroid/blob/7f2ce57c67c68aa921de9231525a4a47ef431fe8/build-logic/convention/src/main/kotlin/com/google/samples/apps/nowinandroid/KotlinAndroid.kt#L63