camera:basic: fail if permissions are not granted
Pre-work for https://github.com/android/ndk-samples/issues/1031. For
some reason the UI for the storage permission isn't shown whenever
targetSdkVersion is 33+, and it's just returning a denial. This sample
is badly structured such that if either permission isn't granted, the UI
init will never run. The two permissions requests really ought to be
decoupled (the viewfinder should still work even if I never try to save
a photo), and the UI should tell the user why it failed. For now I'm
just making it fail more violently, because the lack of any error or
logging is what made this a multi-hour debugging experience instead of
the now obvious problem.