Enable riscv64 for most samples.
This isn't a supported ABI yet, but I want to enable it here to get
ahead of some of that work. I just disabled that config for any of the
samples which don't already work with rv64. There are a few reasons why
some samples don't work with rv64 yet, noted in the comments next to the
line that disables the ABI. Most of them are because some upstream
dependency doesn't yet include rv64 libraries, and there's nothing we
can do about that here.

I haven't actually tested that any of these run because I don't have the
hardware to do so.