Remove san-angeles sample.
This is just another OpenGL sample, and a very old one at that. It
wasn't even originally written for Android, but was ported from asm to
OpenGL for desktop and now this is the Android port (though it still
includes the win32 code!). It's also not doing anything useful that the
other GL samples aren't.